+++ R/select.R
@@ -1,17 +1,23 @@
 #' Select variables using dplyr
 #'
+#' @description
+#'
 #' `step_select()` creates a *specification* of a recipe step that will select
 #' variables using [dplyr::select()].
+#'
+#' `r lifecycle::badge("deprecated")`
 #'
 #' @inheritParams step_center
 #' @param role For model terms selected by this step, what analysis
 #'  role should they be assigned?
 #' @template step-return
-#' @details When an object in the user's global environment is
-#'  referenced in the expression defining the new variable(s),
-#'  it is a good idea to use quasiquotation (e.g. `!!`) to embed
-#'  the value of the object in the expression (to be portable
-#'  between sessions). See the examples.
+#'
+#' @details
+#'
+#' When an object in the user's global environment is referenced in the
+#' expression defining the new variable(s), it is a good idea to use
+#' quasiquotation (e.g. `!!`) to embed the value of the object in the expression
+#' (to be portable between sessions). See the examples.
 #'
 #' # Tidying
 #'
@@ -71,6 +77,11 @@ step_select <- function(
   skip = FALSE,
   id = rand_id("select")
 ) {
+  lifecycle::deprecate_warn(
+    when = "1.3.0",
+    what = "step_select()",
+    details = "See `?select_select()` for recommened alternatives."
+  )
   add_step(
     recipe,
     step_select_new(